+++ changelog.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- Fix for gcr push to encode newlines correctly (#542)
+
 ## v1.0.1429 (2018-11-30)
 
 - Add new property "context" to internal/docker-build (#537)